Disc Filter Cleaning in Drip Irrigation Based on IoT
Disc filter cleaning : Disc filter is to clean the water upto 100 microns. Major disadvantages is the  sedimentation  of sand , dust ,very small particles  etc. So we need to clean the disc filter for two days once by manually for that we can make it to clean by itself in affordable price.